Add MULTI_INSTANCES_SUPPORTED to ShortcutsTile to be able to use any number of Tiles
