+++ common/src/main/java/io/homeassistant/companion/android/common/data/prefs/WearPrefsRepository.kt
@@ -1,11 +1,10 @@
 package io.homeassistant.companion.android.common.data.prefs
 
-import io.homeassistant.companion.android.wear.tiles.ShortcutsTileId
-
 interface WearPrefsRepository {
-    suspend fun getAllTileShortcuts(): Map<ShortcutsTileId, List<String>>
-    suspend fun getTileShortcuts(id: ShortcutsTileId): List<String>
-    suspend fun setTileShortcuts(id: ShortcutsTileId, entities: List<String>)
+    suspend fun getAllTileShortcuts(): Map<out Int?, List<String>>
+    suspend fun getTileShortcuts(tileId: Int): List<String>
+    suspend fun setTileShortcuts(tileId: Int?, entities: List<String>)
+    suspend fun removeTileShortcuts(tileId: Int)
     suspend fun getShowShortcutText(): Boolean
     suspend fun setShowShortcutTextEnabled(enabled: Boolean)
     suspend fun getTemplateTileContent(): String
