--- common/src/main/java/io/homeassistant/companion/android/wear/tiles/ShortcutsTileId.kt
@@ -1,7 +1,0 @@
-package io.homeassistant.companion.android.wear.tiles
-
-enum class ShortcutsTileId {
-    SHORTCUTS_TILE_1,
-    SHORTCUTS_TILE_2,
-    SHORTCUTS_TILE_3
-}
