+++ gradle/libs.versions.toml
@@ -50,7 +50,7 @@ sentry-android = "6.26.0"
 watchfaceComplicationsDataSourceKtx = "1.1.1"
 wear = "1.2.0"
 wear-compose-foundation = "1.1.2"
-wear-tiles = "1.1.0"
+wear-tiles = "1.2.0-alpha05"
 wearPhoneInteractions = "1.0.1"
 wearInput = "1.2.0-alpha02"
 webkit = "1.7.0"
