+++ wear/src/main/java/io/homeassistant/companion/android/home/HomePresenter.kt
@@ -9,7 +9,6 @@ import io.homeassistant.companion.android.common.data.websocket.impl.entities.De
 import io.homeassistant.companion.android.common.data.websocket.impl.entities.EntityRegistryResponse
 import io.homeassistant.companion.android.common.data.websocket.impl.entities.EntityRegistryUpdatedEvent
 import io.homeassistant.companion.android.data.SimplifiedEntity
-import io.homeassistant.companion.android.wear.tiles.ShortcutsTileId
 import kotlinx.coroutines.flow.Flow
 
 interface HomePresenter {
@@ -39,9 +38,9 @@ interface HomePresenter {
     suspend fun getDeviceRegistryUpdates(): Flow<DeviceRegistryUpdatedEvent>?
     suspend fun getEntityRegistryUpdates(): Flow<EntityRegistryUpdatedEvent>?
 
-    suspend fun getAllTileShortcuts(): Map<ShortcutsTileId, List<SimplifiedEntity>>
-    suspend fun getTileShortcuts(id: ShortcutsTileId): List<SimplifiedEntity>
-    suspend fun setTileShortcuts(id: ShortcutsTileId, entities: List<SimplifiedEntity>)
+    suspend fun getAllTileShortcuts(): Map<Int?, List<SimplifiedEntity>>
+    suspend fun getTileShortcuts(tileId: Int): List<SimplifiedEntity>
+    suspend fun setTileShortcuts(tileId: Int?, entities: List<SimplifiedEntity>)
 
     suspend fun getWearHapticFeedback(): Boolean
     suspend fun setWearHapticFeedback(enabled: Boolean)
