+++ wear/src/main/java/io/homeassistant/companion/android/home/HomePresenterImpl.kt
@@ -16,7 +16,6 @@ import io.homeassistant.companion.android.common.data.websocket.impl.entities.En
 import io.homeassistant.companion.android.common.data.websocket.impl.entities.EntityRegistryUpdatedEvent
 import io.homeassistant.companion.android.data.SimplifiedEntity
 import io.homeassistant.companion.android.onboarding.getMessagingToken
-import io.homeassistant.companion.android.wear.tiles.ShortcutsTileId
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.Job
@@ -234,7 +233,7 @@ class HomePresenterImpl @Inject constructor(
         return serverManager.webSocketRepository().getEntityRegistryUpdates()
     }
 
-    override suspend fun getAllTileShortcuts(): Map<ShortcutsTileId, List<SimplifiedEntity>> {
+    override suspend fun getAllTileShortcuts(): Map<Int?, List<SimplifiedEntity>> {
         return wearPrefsRepository.getAllTileShortcuts().mapValues { (_, entities) ->
             entities.map {
                 SimplifiedEntity(it)
@@ -242,12 +241,12 @@ class HomePresenterImpl @Inject constructor(
         }
     }
 
-    override suspend fun getTileShortcuts(id: ShortcutsTileId): List<SimplifiedEntity> {
-        return wearPrefsRepository.getTileShortcuts(id).map { SimplifiedEntity(it) }
+    override suspend fun getTileShortcuts(tileId: Int): List<SimplifiedEntity> {
+        return wearPrefsRepository.getTileShortcuts(tileId).map { SimplifiedEntity(it) }
     }
 
-    override suspend fun setTileShortcuts(id: ShortcutsTileId, entities: List<SimplifiedEntity>) {
-        wearPrefsRepository.setTileShortcuts(id, entities.map { it.entityString })
+    override suspend fun setTileShortcuts(tileId: Int?, entities: List<SimplifiedEntity>) {
+        wearPrefsRepository.setTileShortcuts(tileId, entities.map { it.entityString })
     }
 
     override suspend fun getWearHapticFeedback(): Boolean {
