+++ wear/src/main/java/io/homeassistant/companion/android/onboarding/integration/MobileAppIntegrationPresenterImpl.kt
@@ -8,8 +8,8 @@ import io.homeassistant.companion.android.BuildConfig
 import io.homeassistant.companion.android.common.data.integration.DeviceRegistration
 import io.homeassistant.companion.android.common.data.servers.ServerManager
 import io.homeassistant.companion.android.onboarding.getMessagingToken
-import io.homeassistant.companion.android.tiles.BaseShortcutsTile
 import io.homeassistant.companion.android.tiles.ConversationTile
+import io.homeassistant.companion.android.tiles.ShortcutsTile
 import io.homeassistant.companion.android.tiles.TemplateTile
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
@@ -61,8 +61,8 @@ class MobileAppIntegrationPresenterImpl @Inject constructor(
             val context = view as Context
             val updater = TileService.getUpdater(context)
             updater.requestUpdate(ConversationTile::class.java)
+            updater.requestUpdate(ShortcutsTile::class.java)
             updater.requestUpdate(TemplateTile::class.java)
-            BaseShortcutsTile.requestUpdateForAllShortcutsTiles(context)
         } catch (e: Exception) {
             Log.w(TAG, "Unable to request tiles update")
         }
