--- wear/src/main/java/io/homeassistant/companion/android/tiles/ShortcutsTile3.kt
@@ -1,7 +1,0 @@
-package io.homeassistant.companion.android.tiles
-
-import dagger.hilt.android.AndroidEntryPoint
-import io.homeassistant.companion.android.wear.tiles.ShortcutsTileId
-
-@AndroidEntryPoint
-class ShortcutsTile3 : BaseShortcutsTile(ShortcutsTileId.SHORTCUTS_TILE_3)
